Major UPDATE
I got the auto detect draw working, Now the server will check every minute for the draw date and time, and Draw when time to do so.

added a timed check if player is not online, it will search 5 times for a player online before carrying the pot over to the next draw date.

Still working on injecting the winnings into players bank if they are offline